requirements.txt
Flask
Flask-Mail
opencv-python
numpy
pyzbar
reportlab
pillow